🐛 fix: separated client components and lifted state via context
- Separated client component into a separate file based on server component architecture
- Performed state lifting and created cartContext file to share state between React components
