+++ src/components/features/shop/ProductDetail/CartContext.tsx
@@ -1,0 +1,24 @@
+'use client';
+import React, { createContext, useState, useContext, ReactNode } from 'react';
+
+type CartContextType = {
+  cartCount: number;
+  setCartCount: (count: number) => void;
+};
+
+const CartContext = createContext<CartContextType | undefined>(undefined);
+
+export const useCart = () => {
+  const ctx = useContext(CartContext);
+  if (!ctx) throw new Error('useCart must be used within CartProvider');
+  return ctx;
+};
+
+export function CartProvider({ children }: { children: ReactNode }) {
+  const [cartCount, setCartCount] = useState(0);
+  return (
+    <CartContext.Provider value={{ cartCount, setCartCount }}>
+      {children}
+    </CartContext.Provider>
+  );
+}
